+++ shapes/routenodeshapefactory.h
@@ -11,17 +11,17 @@ namespace Via::Shapes {
 class RouteNodeShapeFactory
 {
 
-    std::unordered_map<char, std::function<RouteNodeShape*(int, int, QColor)>> shapeKeyTable;
+    std::unordered_map<char, std::function<RouteNodeShape*(qreal, qreal, QColor)>> shapeKeyTable;
     std::unordered_map<QString, char> shapeStringKeyTable;
 
     template<typename Shape>
-    RouteNodeShape* generateShape(int x, int y, const QColor &color) {
+    RouteNodeShape* generateShape(qreal x, qreal y, const QColor &color) {
         return new Shape(x, y, color);
     }
 
 public:
     RouteNodeShapeFactory();
-    RouteNodeShape* generateNodeShape(QString key, int x, int y, const QColor &color);
+    RouteNodeShape* generateNodeShape(QString key, qreal x, qreal y, const QColor &color);
 };
 
 }
